Fix issues with rules being run against examples from other rules.
Examples are associated with rules in the configuration files. The test
framework was ignoring this association and was running all rules with
the same RSPEC key against every example. This can lead to tests failing
because they trigger a different rule than the one the example is for.

This change ensures that only the matching rule is executed.